new branch: list-calls
Perl functions take a list of arguments and return a list of return
values; C functions take a list of arguments and return a single
value. But it's sometimes convenient to pretend that C functions return
a list of values, for example by giving them a struct return
type (for example, the gdb-internal function find_pc_sect_line returns a
complicated structure). And, anyway, Platypus isn't just for C. So this
is an attempt to fix that, sacrificing performance where necessary.