new branch: multi, multiple implementations of FFI types
this compiles, but that's only half-way to actually working. One
annoyance that we need to look at again is that the typemap isn't found
unless it's duplicated in lib/, since we're nested too deeply for it to
work otherwise. I'm tempted to skip the ::Impl to save on typing,
anyway.