a PartialType class, for, well, partially-described types.
The point is that if we have a macro definition like

 #define TYPE_INSTANCE_FLAGS(thistype) (thistype)->instance_flags

we know that thistype can only be a pointer to a structure (or union)
containing an instance_flags field. That narrows it down to just a few
possible types, often just one, so we can generate bindings with proper
types.

It's much easier for functions, of course.