start supporting multiple implementations as subobjects of the FFI::Platypus object.
The alternative would be to use inheritance. However, I would like to
allow more fine-grained control over which feature is implemented how
than a single implementation class would provide.

In particular, I want to have access to a symbol resolver based on nm in
addition to the dlsym resolver that's the default.