minor optimization.
The point of this was to make two integer divisions go away. When
allocating runtime-sized arrays on the stack, GCC appears to insist on
issuing a "div" instruction for determining their size, then multiplying
the result by the item size again to find out how much to advance the
stack pointer by.  Using a char array fixes things.