Change the implementation of the arguments accessor API.
Note that the proper Perl way of returning a wide argument is returning
a list of narrow (or at least not-quite-as-wide) replacements. That has
worked for a while, but the old API is there, and I didn't want to break
compatibility.

The new implementation uses a pair of localized global variables, one
for the arguments and one for the types (they're both array refs when in
use), though it just occurs to me we don't set the types yet. Oops. We
fall back to the old implementation when we can't find the localized
global variables, and the old implementation croaks when it can't find
the global arguments variable.

One thing that might or might not be useful is for custom Perl types to
work with variadic functions. The old code would silently overrun the
arguments array, but there's no real reason you shouldn't be able to
implement printf as a custom Perl type with argument_count undef.